requirements.txt
PyYAML==6.0
PyGithub==2.3.0
